feat(portal): inline PaymentSection (invoice + Stripe pay, lazy mount)
